requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.28.1
anthropic==0.49.0
openai==1.58.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-docx==1.1.2
pdfminer.six==20231228
python-dotenv==1.0.1
